refactor: rename & moved files to folders commandsNavigation & commandsBasic
